% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_SE.R
\name{Mean.SE}
\alias{Mean.SE}
\title{Wrapper function that computes Mean and the standard error of the estimate}
\usage{
Mean.SE(data, ..., se.method = "none")
}
\arguments{
\item{data}{data}

\item{\dots}{any other passthru parameters. This include two types of parameters.
The first type is parameters associated with the risk/performance measure, such as tail
probability for VaR and ES. The second type is the parameters associated with the metohd
used to compute the standard error. See \code{\link{SE.IF.iid}}, \code{\link{SE.IF.cor}},
\code{\link{SE.BOOT.iid}}, \code{\link{SE.BOOT.cor}} for details.}

\item{se.method}{a character string indicating which method should be used to compute
the standard error of the estimated standard deviation. One of \code{"none"} (default),
\code{"IFiid"}, \code{"IFcor"}, \code{"BOOTiid"}, \code{"BOOTcor"}.}
}
\value{
A vector or a list depending on se.method
}
\description{
Wrapper function that computes Mean and the standard error of the estimate
}
\author{
Anthony-Alexander Christidis, \email{anthony.christidis@stat.ubc.ca}
}
