% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\name{RachevRatio}
\alias{RachevRatio}
\title{Compute Rachev Ratio}
\usage{
RachevRatio(data, ..., alpha = 0.1, beta = 0.1, rf = 0)
}
\arguments{
\item{data}{vector of data}

\item{...}{other parameters}

\item{alpha}{Lower tail probability.}

\item{beta}{Upper tail probability.}

\item{rf}{Risk-free interest rate.}
}
\value{
Sample Rachev Ratio
}
\description{
Compute Rachev Ratio
}
\examples{
RachevRatio(rnorm(10))
}
\author{
Anthony-Alexander Christidis, \email{anthony.christidis@stat.ubc.ca}
}
