% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSD_SE.R
\name{SemiSD.SE}
\alias{SemiSD.SE}
\title{Standard Error Estimate for Semi-Standared Deviation (SemiSD) of Returns}
\usage{
SemiSD.SE(data, rf = 0, se.method = c("IFiid", "IFcor", "IFcorAdapt",
  "BOOTiid", "BOOTcor", "none")[1:3], prewhiten = FALSE,
  cleanOutliers = FALSE, fitting.method = c("Exponential", "Gamma")[1],
  ...)
}
\arguments{
\item{data}{Data of returns for one or multiple assets or portfolios.}

\item{rf}{Risk-free interest rate.}

\item{se.method}{A character string indicating which method should be used to compute
the standard error of the estimated standard deviation. One or a combination of:
\code{"IFiid"} (default), \code{"IFcor"} (default), \code{"IFcorAdapt"} (default),
\code{"BOOTiid"}, \code{"BOOTcor"}, or \code{"none"}.}

\item{prewhiten}{Boolean variable to indicate if the IF TS is pre-whitened (TRUE) or not (FALSE).}

\item{cleanOutliers}{Boolean variable to indicate whether the pre-whitenning of the influence functions TS should be done through a robust filter.}

\item{fitting.method}{Distribution used in the standard errors computation. Should be one of "Exponential" (default) or "Gamma".}

\item{...}{Additional parameters.}
}
\value{
A vector or a list depending on \code{se.method}.
}
\description{
\code{SemiSD.SE} computes the standard error of the SSD of the returns.
}
\examples{
# Loading data from PerformanceAnalytics
data(edhec, package = "PerformanceAnalytics")
class(edhec)
# Changing the data colnames
names(edhec) = c("CA", "CTA", "DIS", "EM", "EMN",
                 "ED", "FIA", "GM", "LS", "MA",
                 "RV", "SS", "FOF")
# Computing the standard errors for
# the three influence functions based approaches
SemiSD.SE(edhec, se.method=c("IFiid","IFcor","IFcorAdapt"),
          prewhiten=FALSE, cleanOutliers=FALSE,
          fitting.method=c("Exponential", "Gamma")[1])

}
\author{
Anthony-Alexander Christidis, \email{anthony.christidis@stat.ubc.ca}
}
