% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\name{SSD}
\alias{SSD}
\title{Compute Semi-Standard Deviation}
\usage{
SSD(data, ..., rf = 0)
}
\arguments{
\item{data}{vector of data}

\item{...}{other parameters}

\item{rf}{Risk-free interest rate.}
}
\value{
Sample Semi-Standard Deviation
}
\description{
Compute Semi-Standard Deviation
}
\examples{
SSD(rnorm(10))
}
\author{
Anthony-Alexander Christidis, \email{anthony.christidis@stat.ubc.ca}
}
